+++ gr-digital/include/gnuradio/digital/glfsr.h
@@ -23,26 +23,27 @@ namespace digital {
  *
  * \details
  * Generates a maximal length pseudo-random sequence of length 2^degree-1
+ * if given a primitive polynomial.
  */
 class DIGITAL_API glfsr
 {
 private:
-    uint32_t d_shift_register;
-    uint32_t d_mask;
+    uint64_t d_shift_register;
+    uint64_t d_mask;
 
 public:
-    glfsr(uint32_t mask, uint32_t seed)
+    glfsr(uint64_t mask, uint64_t seed)
     {
         d_shift_register = seed;
         d_mask = mask;
     }
     ~glfsr();
 
-    static uint32_t glfsr_mask(unsigned int degree);
+    static uint64_t glfsr_mask(unsigned int degree);
 
     uint8_t next_bit();
 
-    uint32_t mask() const { return d_mask; }
+    uint64_t mask() const { return d_mask; }
 };
 
 } /* namespace digital */
